[refactor][coordinator] Split coordinator source files.
